requirements.txt
watchdog>=4.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
mcp>=1.0.0
pyyaml>=6.0.0
